Python Object Relational Mapping

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,19 +1,24 @@
 #!/usr/bin/python3
-"""Start link class to table in database
+"""
+Deletes all State objects with a name containing
+the letter a from the database hbtn_0e_6_usa.
+Usage: ./13-model_state_delete_a.py <mysql username> /
+                                    <mysql password> /
+                                    <database name>
 """
 import sys
-from model_state import Base, State
-
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
+from model_state import State
 
 if __name__ == "__main__":
-    engine = create_engine('mysql+mysqldb://{}:{}@localhost/{}'.format(
-        sys.argv[1], sys.argv[2], sys.argv[3]), pool_pre_ping=True)
-    Base.metadata.create_all(engine)
+    engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}"
+                           .format(sys.argv[1], sys.argv[2], sys.argv[3]),
+                           pool_pre_ping=True)
     Session = sessionmaker(bind=engine)
     session = Session()
-    states = session.query(State).filter(State.name.like("%a%")).all()
-    for state in states:
-        session.delete(state)
+
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
     session.commit()
